refactor: share cache store adapter
