+++ packages/shared/src/index.ts
@@ -10,3 +10,4 @@ export * from "./cron";
 export * from "./triggers";
 export * from "./completion/extractor";
 export * from "./logger";
+export * from "./cache-store";
