+++ packages/slack-bot/src/branch-preferences.ts
@@ -1,6 +1,6 @@
 import type { Env, SlackInteractionPayload } from "./types";
 import { createLogger } from "./logger";
-import { createKvCacheStore } from "./cache-store";
+import { createKvCacheStore } from "@open-inspect/shared";
 
 const log = createLogger("branch-preferences");
