--- packages/slack-bot/src/cache-store.ts
@@ -1,15 +1,0 @@
-export interface CacheStore {
-  get(key: string): Promise<string | null>;
-  get<T>(key: string, type: "json"): Promise<T | null>;
-  put(key: string, value: string, opts?: KVNamespacePutOptions): Promise<void>;
-  delete(key: string): Promise<void>;
-}
-
-export function createKvCacheStore(kv: KVNamespace): CacheStore {
-  return {
-    get: ((key: string, type?: "json") =>
-      type === "json" ? kv.get(key, "json") : kv.get(key)) as CacheStore["get"],
-    put: (key, value, opts) => (opts ? kv.put(key, value, opts) : kv.put(key, value)),
-    delete: (key) => kv.delete(key),
-  };
-}
